+++ CHANGELOG.md
@@ -7,6 +7,10 @@ mod).
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed the hidden Cheats section in the mod settings to Secrets, since none will be gameplay altering
+
 ## [0.14.19] - 2026-09-24
 
 ### Added